requirements.txt
discord.py==1.6.1
chess==1.4.0
cairosvg==2.5.1
